invert test
